Added Codes I had made to make trippy shapes in matlab
It simulates a gear train and plots points on a certan point in the
mechanism of a cycloid drawing machine (google them to watch some
videos). I kept adding gears to the gear train and playing with the
values of the variabls..There are 9 configuable variables and each
creates a slight variation in the shape drawn. It also creates a list of
points used to plot the graph in a file called "outputx.txt" if you want
to import them in another software.